Fix google nuke

+++ modules/google/nuke.sh
@@ -220,6 +220,21 @@ if [ "$FAIL" -ne 0 ]; then
 fi
 
 PIDS=""
+for line in $(gcloud -q "compute" "addresses" list --uri); do
+  gcloud 'compute' "addresses" delete --project $GOOGLE_PROJECT -q $line &
+  PIDS="$PIDS $!"
+done
+FAIL=0
+for p in $PIDS; do
+  wait $p || let "FAIL+=1"
+  echo $p $FAIL
+done
+if [ "$FAIL" -ne 0 ]; then
+  echo "FAILED to delete addresses. $FAIL"
+  exit 1
+fi
+
+PIDS=""
 for line in $(gcloud -q "compute" "networks" "subnets" list --uri); do
   gcloud 'compute' "networks" "subnets" delete --project $GOOGLE_PROJECT -q $line &
   PIDS="$PIDS $!"
@@ -421,21 +436,6 @@ for p in $PIDS; do
 done
 if [ "$FAIL" -ne 0 ]; then
   echo "FAILED to delete disks. $FAIL"
-  exit 1
-fi
-
-PIDS=""
-for line in $(gcloud -q "compute" "addresses" list --uri); do
-  gcloud 'compute' "addresses" delete --project $GOOGLE_PROJECT -q $line &
-  PIDS="$PIDS $!"
-done
-FAIL=0
-for p in $PIDS; do
-  wait $p || let "FAIL+=1"
-  echo $p $FAIL
-done
-if [ "$FAIL" -ne 0 ]; then
-  echo "FAILED to delete addresses. $FAIL"
   exit 1
 fi
